Release 1.0.1 =============================================================
System Information:
Engine Version: Atom Ver 1.0.1.20201029
Library Version:  Ver 1.0.1.20201029

OS/HW Interface:
Windows Ver 1.0.0.20201013
Linux Ver 1.0.0.20201017

Hosts:
RX Interactive Console Ver 1.0.0.20201029
RX Pipe Child Ver 1.0.0.20201029
RX Simple Host Ver 1.2.0.20201029

Storages:
File System Storage Ver 0.3.5.20201029

Compilers:
MSVC 14.27.29112
GCC 9.3.0

OSes:
Linux Ver:5.4.0-52-generic [x86_64]
Windows Workstation 10.0.18362 [x64]

+++ os_itf/windows/rx_win.c
@@ -1259,7 +1259,7 @@ uint64_t rx_get_us_ticks()
 
 // directories stuff
 
-size_t drives_count = 0;
+int drives_count = 0;
 char** drives = NULL;
 
 void init_fixed_drives()
